edit-logout-added props

+++ admin/src/scripts/Logout.jsx
@@ -1,13 +1,18 @@
 import React, { Component } from 'react'
-// import Form from './home/Form';
+import Form from './home/Form';
 
 export default class Logout extends Component {
-    componentWillMount(){
-        sessionStorage.clear()        
+    state = {msg:''}
+    componentWillMount() {
+        if (sessionStorage.getItem('token')) {
+            sessionStorage.clear()
+            this.setState({msg:'You have been logged out successfully'})
+        }
     }
     render() {
         return (
-            <div>                
+            <div>
+                <Form msg={this.state.msg}></Form>
             </div>
         )
     }
